mcp-client/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40",
    "python-dotenv>=1.0",
    "mcp>=1.0",
    "google-generativeai>=0.8.6",
    "groq>=1.0.0",
]
